Made it so vore compiles on windows. Updated Readme with instructions on how to run it on both

+++ src/vore.hpp
@@ -11,13 +11,13 @@ class Match {
 public:
   std::string value = "";
   std::string replacement = "";
-  u_int64_t file_offset = 0;
-  u_int64_t line_number = 0;
-  u_int64_t match_number = 0;
-  u_int64_t match_length = 0;
+  long long file_offset = 0;
+  long long line_number = 0;
+  long long match_number = 0;
+  long long match_length = 0;
   std::unordered_map<std::string, std::string> variables = std::unordered_map<std::string, std::string>();
 
-  Match(u_int64_t startOffset){
+  Match(long long startOffset){
     file_offset = startOffset;
   };
